Dupliserer pdf-vedlegg-dsl heller enn å trekkje ut til api-model-common
