+++ brevbaker/api-model-common/api/api-model-common.api
@@ -514,18 +514,6 @@ public final class no/nav/pensjon/brevbaker/api/model/EmptyPDFVedleggData : no/n
 	public static final field INSTANCE Lno/nav/pensjon/brevbaker/api/model/EmptyPDFVedleggData;
 }
 
-public final class no/nav/pensjon/brevbaker/api/model/Felt {
-	public fun <init> ()V
-	public final fun add (Ljava/util/Map;)V
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getFelt ()Ljava/util/Map;
-	public fun hashCode ()I
-	public final fun to (Ljava/lang/String;Ljava/lang/Object;)V
-	public final fun to (Ljava/lang/String;Ljava/lang/String;)V
-	public final fun to (Ljava/lang/String;Ljava/util/Map;)V
-	public fun toString ()Ljava/lang/String;
-}
-
 public final class no/nav/pensjon/brevbaker/api/model/LanguageCode : java/lang/Enum {
 	public static final field BOKMAL Lno/nav/pensjon/brevbaker/api/model/LanguageCode;
 	public static final field ENGLISH Lno/nav/pensjon/brevbaker/api/model/LanguageCode;
@@ -826,15 +814,6 @@ public final class no/nav/pensjon/brevbaker/api/model/PDFTittel : no/nav/pensjon
 	public fun toString ()Ljava/lang/String;
 }
 
-public final class no/nav/pensjon/brevbaker/api/model/PDFVedlegg {
-	public fun <init> ()V
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getSider ()Ljava/util/List;
-	public fun hashCode ()I
-	public final fun side (Ljava/lang/String;Lkotlin/jvm/functions/Function1;)V
-	public fun toString ()Ljava/lang/String;
-}
-
 public abstract interface class no/nav/pensjon/brevbaker/api/model/PDFVedleggData {
 }
 
@@ -872,16 +851,6 @@ public final class no/nav/pensjon/brevbaker/api/model/RedigerbareVedleggTitler$V
 	public fun equals (Ljava/lang/Object;)Z
 	public final fun getId-Atnw-NM ()Ljava/lang/String;
 	public final fun getTittel ()Ljava/lang/String;
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
-public final class no/nav/pensjon/brevbaker/api/model/Side {
-	public fun <init> (Ljava/lang/String;)V
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun felt (Lkotlin/jvm/functions/Function1;)V
-	public final fun getFelt ()Ljava/util/List;
-	public final fun getFilnavn ()Ljava/lang/String;
 	public fun hashCode ()I
 	public fun toString ()Ljava/lang/String;
 }
