+++ brevbaker/core/src/main/kotlin/no/nav/brev/brevbaker/PDFVedleggAppender.kt
@@ -1,7 +1,7 @@
 package no.nav.brev.brevbaker
 
+import no.nav.pensjon.brev.template.vedlegg.PDFVedlegg
 import no.nav.pensjon.brevbaker.api.model.LanguageCode
-import no.nav.pensjon.brevbaker.api.model.PDFVedlegg
 
 interface PDFVedleggAppender {
     fun leggPaaVedlegg(
