+++ pensjon/brevbaker/src/main/kotlin/no/nav/pensjon/brev/pdfvedlegg/VedleggAppender.kt
@@ -1,7 +1,7 @@
 package no.nav.pensjon.brev.pdfvedlegg
 
+import no.nav.pensjon.brev.template.vedlegg.PDFVedlegg
 import no.nav.pensjon.brevbaker.api.model.LanguageCode
-import no.nav.pensjon.brevbaker.api.model.PDFVedlegg
 import org.apache.pdfbox.Loader
 import org.apache.pdfbox.cos.COSName
 import org.apache.pdfbox.multipdf.PDFMergerUtility
